Comparison of Total Sales Between Dry and Rainy Seasons: A Statistical Analysis
The data compares total sales between two seasons—Dry and Rainy—across several days. In the Dry Season, the average sales value is higher (NGN 42,779.81) with more variation (standard deviation of 15,880.13). The Rainy Season has a lower average sales value (NGN 35,929.95), but with a smaller variation (standard deviation of 14,353.98).

A statistical test (T-test) was conducted to compare total sales between the two seasons. The result showed a T-statistic of 14.97 and a P-value of nearly zero (1.25e-49). Since the P-value is extremely low, we reject the null hypothesis, indicating that the total sales between the Dry and Rainy seasons are statistically significantly different.